Rename leshan-bs-server in leshan-bsserver-demo.
