+++ leshan-bsserver-demo/src/main/java/org/eclipse/leshan/server/bootstrap/demo/BootstrapSecurityStore.java
@@ -13,7 +13,7 @@
  * Contributors:
  *     Sierra Wireless - initial API and implementation
  *******************************************************************************/
-package org.eclipse.leshan.bootstrap;
+package org.eclipse.leshan.server.bootstrap.demo;
 
 import java.util.Arrays;
 import java.util.Map;
