+++ leshan-bsserver-demo/src/main/resources/log4j2.xml
@@ -16,7 +16,7 @@ Contributors:
     Sierra Wireless - initial API and implementation
 
  -->
-<Configuration status="warn" strict="true" name="leshan standalone"
+<Configuration status="warn" strict="true" name="leshan bootstrap server demo"
                packages="org.apache.logging.log4j.test">
 
   <Appenders>
