+++ leshan-bsserver-demo/src/main/resources/webapp/img/fav_icon.png
