+++ leshan-bsserver-demo/src/main/resources/webapp/img/logo.png
