+++ leshan-bsserver-demo/src/main/resources/webapp/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
     <head>
         <meta charset="utf-8">
-        <title>Leshan Bootstrap</title>
+        <title>Leshan Bootstrap Demo</title>
 
         <link rel="stylesheet" href="vendor/css/bootstrap.min.css">
         <link rel="stylesheet" href="css/lw-app.css" />
