+++ leshan-bsserver-demo/src/main/resources/webapp/js/bsconfigstore.js
