+++ leshan-bsserver-demo/src/main/resources/webapp/tag/bootstrap.tag
